Inherit discriminator from pojos
- The discriminator for a oneOf
composition might not necessarily
be on the same level as the oneOf
composition is defined. It might
also be in parent schemas used
by all schemas in the composition
via allOf composition. A
discriminator is now used also
if the same is found in each
schema withing the composition.

Issue: 136