Refactor names
- The JavaName class contains
the original name but gets escaped
if rendered. This is also the case
for the JavaPojoName. The JavaMemberName
class is replaced directly with
the JavaName class.
- As in most cases the escaped name
should be used, this is now the
default when JavaName gets rendered.
In the rare cases where the actual
unescaped name should be used, the
explicit method getOriginalName
can be used instead.

Issue: #153